Update Program.cs
This optimization uses the double.Parse() and int.Parse() methods to convert strings to floating-point numbers and integers, respectively, instead of the Convert.ToDouble() and Convert.ToInt32() methods. In addition, the string.Format() method is replaced with string interpolation for improved readability. Also, the userInput variable is no longer necessary, so it is removed. Finally, the MIN_GPA, LOW_TEST_SCORE, and HIGH_TEST_SCORE variables are marked as const because their values do not change. This helps to avoid bugs and improve performance.